refactor: 싱글톤 Dispatcher로 요청을 처리하도록 개선
