+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -6,16 +6,20 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.handler.HttpRequestHandler;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class Http11Processor implements Runnable, Processor {
 
     private static final Logger log = LoggerFactory.getLogger(Http11Processor.class);
-
+    private final Dispatcher dispatcher;
     private final Socket connection;
 
     public Http11Processor(final Socket connection) {
+        dispatcher = Dispatcher.getInstance();
         this.connection = connection;
     }
 
@@ -28,13 +32,13 @@ public class Http11Processor implements Runnable, Processor {
     @Override
     public void process(final Socket connection) {
         try (final var inputStream = connection.getInputStream();
-             final var outputStream = connection.getOutputStream()) {
+             final var outputStream = connection.getOutputStream();
+             final BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream))) {
 
-            final BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
             HttpRequest httpRequest = HttpRequest.from(bufferedReader.readLine());
-            StaticResourceHandler staticResourceHandler = new StaticResourceHandler();
+            HttpRequestHandler httpRequestHandler = dispatcher.mappedHandler(httpRequest);
 
-            String responseBody = staticResourceHandler.handle(httpRequest);
+            String responseBody = httpRequestHandler.handle(httpRequest);
             HttpResponse httpResponse = HttpResponse.htmlResourceOkResponse(outputStream, responseBody);
             httpResponse.flush();
         } catch (IOException | UncheckedServletException e) {
