+++ tomcat/src/main/java/org/apache/coyote/http11/handler/HttpRequestHandler.java
@@ -1,6 +1,7 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.handler;
 
 import java.io.IOException;
+import org.apache.coyote.http11.request.HttpRequest;
 
 public interface HttpRequestHandler {
