+++ tomcat/src/main/java/org/apache/coyote/http11/request/Method.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import com.techcourse.exception.UncheckedServletException;
 import java.util.Arrays;
